+++ lib/easypost/services/user.rb
@@ -78,4 +78,21 @@ Please use `retrieve_api_keys_for_user` in the `api_key` service instead.'
 
     EasyPost::InternalUtilities::Json.convert_json_to_object(response, EasyPost::Models::Brand)
   end
+
+  # Retrieve all child Users.
+  def all_children(params = {})
+    filters = { key: 'children' }
+
+    get_all_helper('users/children', EasyPost::Models::User, params, filters)
+  end
+
+  # Get the next page of child users.
+  def get_next_page_of_children(collection, page_size = nil)
+    raise EasyPost::Errors::EndOfPaginationError.new unless more_pages?(collection)
+
+    params = { before_id: collection.children.last.id }
+    params[:page_size] = page_size unless page_size.nil?
+
+    all_children(params)
+  end
 end